Move connect and quit commands to command package.
Also renames cmds_misc.go to cmd_cluster_id.go because that's the only
command remaining in the file.
