+++ pkg/client/cli/connect/cmd.go
@@ -1,58 +1,38 @@
-package cli
+package connect
 
 import (
 	"fmt"
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 	empty "google.golang.org/protobuf/types/known/emptypb"
-	"k8s.io/cli-runtime/pkg/genericclioptions"
-	"k8s.io/client-go/kubernetes"
 
-	"github.com/datawire/k8sapi/pkg/k8sapi"
 	rpc "github.com/telepresenceio/telepresence/rpc/v2/daemon"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/ann"
-	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/connect"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/daemon"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/socket"
 )
 
-// ClusterIdCommand is a simple command that makes it easier for users to
-// figure out what their cluster ID is. For now this is just used when
-// people are making licenses for air-gapped environments.
-func ClusterIdCommand() *cobra.Command {
-	kubeConfig := genericclioptions.NewConfigFlags(false)
-	cmd := &cobra.Command{
-		Use:  "current-cluster-id",
-		Args: cobra.NoArgs,
+func Command() *cobra.Command {
+	var request *daemon.Request
 
-		Short: "Get cluster ID for your kubernetes cluster",
-		Long:  "Get cluster ID for your kubernetes cluster, mostly used for licenses in air-gapped environments",
-		RunE: func(cmd *cobra.Command, _ []string) error {
-			restConfig, err := kubeConfig.ToRESTConfig()
-			if err != nil {
-				return err
-			}
-			ki, err := kubernetes.NewForConfig(restConfig)
-			if err != nil {
-				return err
-			}
-			clusterID, err := k8sapi.GetClusterID(k8sapi.WithK8sInterface(cmd.Context(), ki))
-			if err != nil {
-				return err
-			}
-			fmt.Fprintf(cmd.OutOrStdout(), "Cluster ID: %s\n", clusterID)
-			return nil
+	cmd := &cobra.Command{
+		Use:   "connect [flags] [-- <command to run while connected>]",
+		Args:  cobra.ArbitraryArgs,
+		Short: "Connect to a cluster",
+		Annotations: map[string]string{
+			ann.Session: ann.Required,
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			request.CommitFlags(cmd)
+			return RunConnect(cmd, args)
 		},
 	}
-	kubeFlags := pflag.NewFlagSet("Kubernetes flags", 0)
-	kubeConfig.AddFlags(kubeFlags)
-	cmd.Flags().AddFlagSet(kubeFlags)
+	request = daemon.InitRequest(cmd)
 	return cmd
 }
 
-func quitCommand() *cobra.Command {
+func QuitCommand() *cobra.Command {
 	quitDaemons := false
 	quitRootDaemon := false
 	quitUserDaemon := false
@@ -63,7 +43,7 @@ func quitCommand() *cobra.Command {
 		Short:       "Tell telepresence daemon to quit",
 		Annotations: map[string]string{ann.UserDaemon: ann.Optional},
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			if err := connect.InitCommand(cmd); err != nil {
+			if err := InitCommand(cmd); err != nil {
 				return err
 			}
 			if quitUserDaemon {
@@ -82,7 +62,7 @@ func quitCommand() *cobra.Command {
 					_, _ = rpc.NewDaemonClient(conn).Quit(ctx, &empty.Empty{})
 				}
 			}
-			return connect.Disconnect(cmd.Context(), quitDaemons)
+			return Disconnect(cmd.Context(), quitDaemons)
 		},
 	}
 	flags := cmd.Flags()
